+++ PrettyDocComments/Model/Comment.cs
@@ -1,4 +1,7 @@
-﻿using Microsoft.VisualStudio.Text;
+﻿using System.Windows.Forms;
+using System.Windows.Shapes;
+using Microsoft.VisualStudio.Text;
+using Microsoft.VisualStudio.Text.Editor;
 
 namespace PrettyDocComments.Model;
 
@@ -33,5 +36,19 @@ internal readonly struct Comment<TData>
     public Comment<TNewData> ConvertTo<TNewData>(TNewData newData)
     {
         return new Comment<TNewData>(Span, CommentLeftCharIndex, FirstLineNumber, LastLineNumber, Width, newData);
+    }
+
+    public bool ContainsCaretOrSelStartOrEnd(IWpfTextView view)
+    {
+        int caretLineNumber = view.Caret?.Position.BufferPosition.GetContainingLineNumber() ?? -1;
+        if (ContainsLine(caretLineNumber)) {
+            return true;
+        }
+        if (view.Selection is { } selection) {
+            return
+                ContainsLine(selection.Start.Position.GetContainingLine().LineNumber) ||
+                ContainsLine(selection.End.Position.GetContainingLine().LineNumber);
+        }
+        return false;
     }
 }